+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Significantly improved the color palette used by the Terrain Scanner app.
 - Made the refueling ETA more accurate.
 - Added some mild inaccuracy to the stability bar.
 - Updated interior door texture slightly to more easily distinguish between the front and sides.